rename: 파일 구조 변경
- external -> external/S3
